refactor(assistant): move turn/watchdog telemetry stores out of memory/ into telemetry/
