+++ assistant/src/telemetry/__tests__/watchdog-events-store.test.ts
