+++ assistant/src/telemetry/watchdog-events-store.ts
